[Api] Move ContextBuilder to Serializer/ContextBuilder directory
