+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/context_builders.xml
@@ -19,7 +19,8 @@
         <defaults public="false" />
 
         <service
-            id="Sylius\Bundle\ApiBundle\SerializerContextBuilder\ChannelCodeAwareContextBuilder"
+            id="sylius_api.context_builder.channel_code_aware"
+            class="Sylius\Bundle\ApiBundle\Serializer\ContextBuilder\ChannelCodeAwareContextBuilder"
             decorates="api_platform.serializer.context_builder"
             decoration-priority="64"
         >
@@ -30,7 +31,8 @@
         </service>
 
         <service
-            id="Sylius\Bundle\ApiBundle\SerializerContextBuilder\LoggedInCustomerEmailAwareContextBuilder"
+            id="sylius_api.context_builder.logged_in_customer_email_aware"
+            class="Sylius\Bundle\ApiBundle\Serializer\ContextBuilder\LoggedInCustomerEmailAwareContextBuilder"
             decorates="api_platform.serializer.context_builder"
             decoration-priority="64"
         >
@@ -50,7 +52,8 @@
         </service>
 
         <service
-            id="Sylius\Bundle\ApiBundle\SerializerContextBuilder\LocaleCodeAwareContextBuilder"
+            id="sylius_api.context_builder.locale_code_aware"
+            class="Sylius\Bundle\ApiBundle\Serializer\ContextBuilder\LocaleCodeAwareContextBuilder"
             decorates="api_platform.serializer.context_builder"
             decoration-priority="64"
         >
@@ -78,8 +81,8 @@
         </service>
 
         <service
-            id="sylius_api.serializer_context_builder.logged_in_shop_user_id_aware"
-            class="Sylius\Bundle\ApiBundle\SerializerContextBuilder\LoggedInShopUserIdAwareContextBuilder"
+            id="sylius_api.context_builder.logged_in_shop_user_id_aware"
+            class="Sylius\Bundle\ApiBundle\Serializer\ContextBuilder\LoggedInShopUserIdAwareContextBuilder"
             decorates="api_platform.serializer.context_builder"
             decoration-priority="64"
         >
@@ -90,8 +93,8 @@
         </service>
 
         <service
-            id="sylius_api.serializer_context_builder.shipment_aware"
-            class="Sylius\Bundle\ApiBundle\SerializerContextBuilder\UriVariablesAwareContextBuilder"
+            id="sylius_api.context_builder.shipment_aware"
+            class="Sylius\Bundle\ApiBundle\Serializer\ContextBuilder\UriVariablesAwareContextBuilder"
             decorates="api_platform.serializer.context_builder"
             decoration-priority="64"
         >
@@ -103,8 +106,8 @@
         </service>
 
         <service
-            id="sylius_api.serializer_context_builder.order_token_value_aware"
-            class="Sylius\Bundle\ApiBundle\SerializerContextBuilder\UriVariablesAwareContextBuilder"
+            id="sylius_api.context_builder.order_token_value_aware"
+            class="Sylius\Bundle\ApiBundle\Serializer\ContextBuilder\UriVariablesAwareContextBuilder"
             decorates="api_platform.serializer.context_builder"
             decoration-priority="64"
         >
@@ -116,8 +119,8 @@
         </service>
 
         <service
-            id="sylius_api.serializer_context_builder.order_item_aware"
-            class="Sylius\Bundle\ApiBundle\SerializerContextBuilder\UriVariablesAwareContextBuilder"
+            id="sylius_api.context_builder.order_item_aware"
+            class="Sylius\Bundle\ApiBundle\Serializer\ContextBuilder\UriVariablesAwareContextBuilder"
             decorates="api_platform.serializer.context_builder"
             decoration-priority="64"
         >
