+++ src/Sylius/Bundle/ApiBundle/Serializer/ContextBuilder/LoggedInCustomerEmailAwareContextBuilder.php
@@ -11,15 +11,18 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\SerializerContextBuilder;
+namespace Sylius\Bundle\ApiBundle\Serializer\ContextBuilder;
 
 use ApiPlatform\Serializer\SerializerContextBuilderInterface;
-use Sylius\Bundle\ApiBundle\Command\ShopUserIdAwareInterface;
+use Sylius\Bundle\ApiBundle\Command\LoggedInCustomerEmailAwareInterface;
 use Sylius\Bundle\ApiBundle\Context\UserContextInterface;
+use Sylius\Component\Core\Model\CustomerInterface;
 use Sylius\Component\Core\Model\ShopUserInterface;
+use Sylius\Component\User\Model\UserInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Webmozart\Assert\Assert;
 
-final class LoggedInShopUserIdAwareContextBuilder extends AbstractInputContextBuilder
+final class LoggedInCustomerEmailAwareContextBuilder extends AbstractInputContextBuilder
 {
     public function __construct(
         SerializerContextBuilderInterface $decoratedContextBuilder,
@@ -32,23 +35,31 @@ final class LoggedInShopUserIdAwareContextBuilder extends AbstractInputContextBu
 
     protected function supportsClass(string $class): bool
     {
-        return is_a($class, ShopUserIdAwareInterface::class, true);
+        return is_a($class, LoggedInCustomerEmailAwareInterface::class, true);
     }
 
     protected function supports(Request $request, array $context, ?array $extractedAttributes): bool
     {
-        return $this->getShopUser() !== null;
+        return !isset($request->toArray()['email']) &&
+            $this->getCustomer() !== null;
     }
 
     protected function resolveValue(array $context, ?array $extractedAttributes): mixed
     {
-        return $this->getShopUser()->getId();
+        return $this->getCustomer()->getEmail();
     }
 
-    private function getShopUser(): ?ShopUserInterface
+    private function getCustomer(): ?CustomerInterface
     {
+        /** @var UserInterface|null $user */
         $user = $this->userContext->getUser();
+        if ($user instanceof ShopUserInterface) {
+            $customer = $user->getCustomer();
+            Assert::nullOrIsInstanceOf($customer, CustomerInterface::class);
 
-        return $user instanceof ShopUserInterface ? $user : null;
+            return $customer;
+        }
+
+        return null;
     }
 }
