+++ src/Sylius/Bundle/ApiBundle/Serializer/ContextBuilder/LoggedInShopUserIdAwareContextBuilder.php
@@ -11,18 +11,15 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\SerializerContextBuilder;
+namespace Sylius\Bundle\ApiBundle\Serializer\ContextBuilder;
 
 use ApiPlatform\Serializer\SerializerContextBuilderInterface;
-use Sylius\Bundle\ApiBundle\Command\LoggedInCustomerEmailAwareInterface;
+use Sylius\Bundle\ApiBundle\Command\ShopUserIdAwareInterface;
 use Sylius\Bundle\ApiBundle\Context\UserContextInterface;
-use Sylius\Component\Core\Model\CustomerInterface;
 use Sylius\Component\Core\Model\ShopUserInterface;
-use Sylius\Component\User\Model\UserInterface;
 use Symfony\Component\HttpFoundation\Request;
-use Webmozart\Assert\Assert;
 
-final class LoggedInCustomerEmailAwareContextBuilder extends AbstractInputContextBuilder
+final class LoggedInShopUserIdAwareContextBuilder extends AbstractInputContextBuilder
 {
     public function __construct(
         SerializerContextBuilderInterface $decoratedContextBuilder,
@@ -35,31 +32,23 @@ final class LoggedInCustomerEmailAwareContextBuilder extends AbstractInputContex
 
     protected function supportsClass(string $class): bool
     {
-        return is_a($class, LoggedInCustomerEmailAwareInterface::class, true);
+        return is_a($class, ShopUserIdAwareInterface::class, true);
     }
 
     protected function supports(Request $request, array $context, ?array $extractedAttributes): bool
     {
-        return !isset($request->toArray()['email']) &&
-            $this->getCustomer() !== null;
+        return $this->getShopUser() !== null;
     }
 
     protected function resolveValue(array $context, ?array $extractedAttributes): mixed
     {
-        return $this->getCustomer()->getEmail();
+        return $this->getShopUser()->getId();
     }
 
-    private function getCustomer(): ?CustomerInterface
+    private function getShopUser(): ?ShopUserInterface
     {
-        /** @var UserInterface|null $user */
         $user = $this->userContext->getUser();
-        if ($user instanceof ShopUserInterface) {
-            $customer = $user->getCustomer();
-            Assert::nullOrIsInstanceOf($customer, CustomerInterface::class);
 
-            return $customer;
-        }
-
-        return null;
+        return $user instanceof ShopUserInterface ? $user : null;
     }
 }
