+++ src/Sylius/Bundle/ApiBundle/spec/Serializer/ContextBuilder/LocaleCodeAwareContextBuilderSpec.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Bundle\ApiBundle\SerializerContextBuilder;
+namespace spec\Sylius\Bundle\ApiBundle\Serializer\ContextBuilder;
 
 use ApiPlatform\Serializer\SerializerContextBuilderInterface;
 use PhpSpec\ObjectBehavior;
